Use common page structure for the main pages
