+++ apps/web/app/(app)/[emailAccountId]/bulk-unsubscribe/BulkActions.tsx
@@ -68,7 +68,7 @@ export function BulkActions({
   return (
     <>
       <PremiumTooltip showTooltip={!hasUnsubscribeAccess} openModal={openModal}>
-        <div className="flex items-center space-x-1.5">
+        <div className="flex items-center space-x-1.5 mt-2">
           <div>
             <Button
               size="sm"
