+++ apps/web/app/(app)/[emailAccountId]/stats/Stats.tsx
@@ -15,7 +15,8 @@ import { EmailActionsAnalytics } from "@/app/(app)/[emailAccountId]/stats/EmailA
 import { BulkUnsubscribeSummary } from "@/app/(app)/[emailAccountId]/bulk-unsubscribe/BulkUnsubscribeSummary";
 import { CardBasic } from "@/components/ui/card";
 import { Title } from "@tremor/react";
-import { TopBar } from "@/components/TopBar";
+import { PageHeading } from "@/components/Typography";
+import { PageWrapper } from "@/components/PageWrapper";
 
 const selectOptions = [
   { label: "Last week", value: "7" },
@@ -64,8 +65,9 @@ export function Stats() {
   }, [onLoad]);
 
   return (
-    <div className="pb-20">
-      <TopBar sticky>
+    <PageWrapper className="pb-20">
+      <PageHeading>Analytics</PageHeading>
+      <div className="flex items-center justify-between">
         {isLoading ? <LoadProgress /> : <div />}
         <div className="flex flex-wrap gap-1">
           <ActionBar
@@ -80,20 +82,15 @@ export function Stats() {
           />
           <LoadStatsButton />
         </div>
-      </TopBar>
+      </div>
 
-      <div className="px-4 py-4">
+      <div className="py-4">
         <StatsSummary dateRange={dateRange} refreshInterval={refreshInterval} />
       </div>
 
-      <div className="px-4">
-        <EmailAnalytics
-          dateRange={dateRange}
-          refreshInterval={refreshInterval}
-        />
-      </div>
+      <EmailAnalytics dateRange={dateRange} refreshInterval={refreshInterval} />
 
-      <div className="mx-4 mt-4">
+      <div className="mt-4">
         <DetailedStats
           dateRange={dateRange}
           period={period}
@@ -101,7 +98,7 @@ export function Stats() {
         />
       </div>
 
-      <div className="mt-4 px-4">
+      <div className="mt-4">
         <CardBasic>
           <Title>
             How many emailers you've handled with Inbox Zero bulk unsubscribe
@@ -112,11 +109,11 @@ export function Stats() {
         </CardBasic>
       </div>
 
-      <div className="mt-4 px-4">
+      <div className="mt-4">
         <EmailActionsAnalytics />
       </div>
 
       <StatsOnboarding />
-    </div>
+    </PageWrapper>
   );
 }
