+++ apps/web/components/PageWrapper.tsx
@@ -1,0 +1,15 @@
+import { cn } from "@/utils";
+
+export function PageWrapper({
+  children,
+  className,
+}: {
+  children: React.ReactNode;
+  className?: string;
+}) {
+  return (
+    <div className={cn("mx-auto max-w-screen-xl w-full", className)}>
+      {children}
+    </div>
+  );
+}
